@@ -13,26 +13,27 @@
     "babel-jest": "*",
     "babel-preset-es2015": "^6.14.0",
     "deep-equal": "^1.0.1",
-    "gulp": "~3.9.0",
+    "gulp": "^3.9.1",
     "gulp-babel": "^6.1.2",
     "gulp-clean": "^0.3.2",
     "gulp-collect": "^0.1.0",
     "gulp-jest": "^0.5.2",
     "gulp-rename": "^1.2.2",
-    "gulp-tslint": "^6.0.1",
-    "gulp-typescript": "^2.13.6",
+    "gulp-tslint": "^8.1.2",
+    "gulp-typescript": "^3.2.3",
     "gulp-util": "^3.0.7",
     "jest-cli": "^15.1.1",
     "lodash": "^3.10.0",
     "q": "^1.4.1",
     "recast": "^0.11.10",
     "through2": "^2.0.1",
-    "tslint": "^3.13.0",
-    "typescript": "^2.0.0",
-    "typings": "^1.3.0",
+    "tslint": "^5.8.0",
+    "typescript": "^2.5.3",
+    "typings": "^2.1.1",
     "vm": "^0.1.0"
   },
   "dependencies": {
+    "glob-stream": "^6.1.0",
     "minimist": "^1.2.0"
   },
   "jest": {
