@@ -1,20 +1,20 @@
-import Process = require("../../kernel/kernel/process");
 import {
-    addProcess, processTable, storeProcessTable,
-    getProcessById, sleepProcess, killProcess
+    addProcess, getProcessById, killProcess,
+    processTable, sleepProcess, storeProcessTable,
 } from "../../kernel/kernel/kernel";
-import SpawnProcess = require("./spawn");
-import LibrarianProcess = require("./librarian");
-import UpgraderProcess = require("./upgrader");
-import DefenseProcess = require("./defense");
-import MaintainerProcess = require("./maintainer");
+import Process = require("../../kernel/kernel/process");
 import MiningProcess = require("../mining/mining");
-import StarterProcess = require("./starter");
+import DefenseProcess = require("./defense");
+import LibrarianProcess = require("./librarian");
 import LinkManagerProcess = require("./link-manager");
+import MaintainerProcess = require("./maintainer");
+import SpawnProcess = require("./spawn");
+import StarterProcess = require("./starter");
+import UpgraderProcess = require("./upgrader");
 // import BuilderPlannerProcess = require("./building-planner");
 class ColonyProcess extends Process {
     public static start(roomName: string) {
-        let p = new ColonyProcess(0, 0);
+        const p = new ColonyProcess(0, 0);
         addProcess(p);
         storeProcessTable();
 
@@ -36,41 +36,41 @@ class ColonyProcess extends Process {
     }
 
     public run(): number {
-        let memory = this.memory;
-        let room = Game.rooms[memory.roomName];
+        const memory = this.memory;
+        const room = Game.rooms[memory.roomName];
 
-        let spawnPID = memory.spawnPID;
+        const spawnPID = memory.spawnPID;
         if (!spawnPID || !getProcessById(spawnPID)) {
             memory.spawnPID = this.launchSpawnProcess(room.name);
         }
         if (room.controller!.level >= 4 && room.storage && room.storage.store.energy > 10000) {
 
-            let upgraderPID = memory.upgraderPID;
+            const upgraderPID = memory.upgraderPID;
 
             if (!upgraderPID || !getProcessById(upgraderPID)) {
                 console.log("Starting upgrader process for Room:" + room.name);
                 memory.upgraderPID = UpgraderProcess.start(memory.roomName, this.pid);
             }
 
-            let defenderPID = memory.defenderPID;
+            const defenderPID = memory.defenderPID;
             if (!defenderPID || !getProcessById(defenderPID)) {
                 console.log("Starting defender process for Room:" + room.name);
                 memory.defenderPID = DefenseProcess.start(memory.roomName, this.pid);
             }
 
-            let maintainerPID = memory.maintainerPID;
+            const maintainerPID = memory.maintainerPID;
             if (!maintainerPID || !getProcessById(maintainerPID)) {
                 console.log("Starting maintainer process for room:" + room.name);
                 memory.maintainerPID = MaintainerProcess.start(memory.roomName, this.pid);
             }
 
-            let linkPID = memory.linkPID;
+            const linkPID = memory.linkPID;
             if (!linkPID || !getProcessById(linkPID)) {
                 console.log("Starting link process for room:" + room.name);
                 memory.linkPID = LinkManagerProcess.start(memory.roomName, this.pid);
             }
 
-            let librarianPID = memory.librarianPID;
+            const librarianPID = memory.librarianPID;
             if (!librarianPID || !getProcessById(librarianPID)) {
                 const p = new LibrarianProcess(0, this.pid);
                 addProcess(p);
@@ -79,13 +79,13 @@ class ColonyProcess extends Process {
             }
 
             let inRoomMiningPid: number[] = memory.miningPIDList || [];
-            inRoomMiningPid = _.filter(inRoomMiningPid, pid => getProcessById(pid));
+            inRoomMiningPid = _.filter(inRoomMiningPid, (pid) => getProcessById(pid));
             if (inRoomMiningPid.length === 0) {
                 const sources = room.find(FIND_SOURCES) as Source[];
-                for (let source of sources) {
+                for (const source of sources) {
                     const sourceId = source.id;
                     const roomName = room.name;
-                    let flag = Game.flags[sourceId];
+                    const flag = Game.flags[sourceId];
                     if (!flag) {
                         source.pos.createFlag(sourceId, COLOR_YELLOW);
                     }
@@ -103,7 +103,6 @@ class ColonyProcess extends Process {
             sleepProcess(this, 100);
         }
 
-
         return 0;
     }
 
@@ -119,7 +118,7 @@ class ColonyProcess extends Process {
 
     private launchSpawnProcess(roomName: string) {
         console.log("Starting spawn process for room:" + roomName);
-        let p = SpawnProcess.start(roomName, this.pid);
+        const p = SpawnProcess.start(roomName, this.pid);
         p.parentPID = this.pid;
         return p.pid;
     }
