@@ -1,5 +1,5 @@
-import Process = require("../../kernel/kernel/process");
 import { getProcessById } from "../../kernel/kernel/kernel";
+import Process = require("../../kernel/kernel/process");
 import { getSpawnProcess } from "../../utils/spawn";
 class LibrarianProcess extends Process {
 
@@ -16,7 +16,7 @@ class LibrarianProcess extends Process {
         }
     }
     public runCreep(creepName: string): number {
-        let creep = Game.creeps[creepName];
+        const creep = Game.creeps[creepName];
 
         const storage = creep.room.storage;
         let toSpawn = true;
@@ -32,24 +32,24 @@ class LibrarianProcess extends Process {
                 }
             }
         } else {
-            let storages: (Extension | Spawn)[] = <(Extension | Spawn)[]>creep.room.find(FIND_MY_STRUCTURES, {
-                filter: function (s: any) {
+            let storages: Array<Extension | Spawn> = creep.room.find(FIND_MY_STRUCTURES, {
+                filter(s: any) {
                     return (s.structureType === STRUCTURE_EXTENSION || s.structureType === STRUCTURE_TOWER)
                         && s.needEnergy();
                 },
-            });
-            let spawnInDestRoom = creep.room.find(FIND_MY_SPAWNS, { filter: function (spawn: Spawn) { return spawn.energy < spawn.energyCapacity; } });
+            }) as Array<Extension | Spawn>;
+            const spawnInDestRoom = creep.room.find(FIND_MY_SPAWNS, { filter(spawn: Spawn) { return spawn.energy < spawn.energyCapacity; } });
             if (toSpawn)
                 storages.push.apply(storages, spawnInDestRoom);
 
             if (storages.length) {
-                let min = _.min(storages, function (item) {
+                let min = _.min(storages, function(item) {
                     return creep.pos.getRangeTo(item);
                 });
                 if (creep.pos.isNearTo(min)) {
                     creep.transfer(min, RESOURCE_ENERGY);
-                    storages = _.filter(storages, x => x.id !== min.id);
-                    min = _.min(storages, function (item) {
+                    storages = _.filter(storages, (x) => x.id !== min.id);
+                    min = _.min(storages, function(item) {
                         return creep.pos.getRangeTo(item);
                     });
                     creep.moveTo(min);
@@ -65,8 +65,8 @@ class LibrarianProcess extends Process {
     }
 
     public spawnSmallCreep(): number {
-        let room = Game.rooms[this.memory.roomName];
-        let availableEnergy = room.energyAvailable;
+        const room = Game.rooms[this.memory.roomName];
+        const availableEnergy = room.energyAvailable;
         let modifier = Math.floor(availableEnergy / 100);
         if (modifier > 25)
             modifier = 25;
@@ -76,8 +76,8 @@ class LibrarianProcess extends Process {
     }
 
     public spawnNormalCreep(): number {
-        let room = Game.rooms[this.memory.roomName];
-        let availableEnergy = room.energyCapacityAvailable;
+        const room = Game.rooms[this.memory.roomName];
+        const availableEnergy = room.energyCapacityAvailable;
         let modifier = Math.floor(availableEnergy / 100);
         if (modifier > 25)
             modifier = 25;
@@ -87,11 +87,11 @@ class LibrarianProcess extends Process {
     }
 
     public run(): number {
-        let colonyProcess = getProcessById(this.parentPID);
+        const colonyProcess = getProcessById(this.parentPID);
         if (!colonyProcess)
             return this.stop(0);
 
-        let memory = this.memory;
+        const memory = this.memory;
         let smallCreepName = memory.smallCreepName;
         let creepName = memory.creepName;
 
@@ -118,7 +118,7 @@ class LibrarianProcess extends Process {
     }
 
     protected spawnCreep(creepID: string, bodyParts: bodyMap, priority?: number) {
-        let spawnProcess = getSpawnProcess(this.memory.roomName);
+        const spawnProcess = getSpawnProcess(this.memory.roomName);
 
         if (spawnProcess) {
             spawnProcess.spawn(creepID, bodyParts, this.pid, priority);
