@@ -1,5 +1,5 @@
+import { getProcessById, sleepProcess } from "../../kernel/kernel/kernel"
 import Process = require("../../kernel/kernel/process");
-import { sleepProcess, getProcessById } from "../../kernel/kernel/kernel"
 import ColonyProcess = require("./colony");
 
 class RoadPlannerProcess extends Process {
@@ -14,9 +14,9 @@ class RoadPlannerProcess extends Process {
         const room = Game.rooms[colonyProcess.memory.roomName];
         const controller = room.controller!;
         const sources = room.find(FIND_SOURCES) as Source[];
-        for (let source of sources) {
+        for (const source of sources) {
             const path = PathFinder.search(source.pos, controller.pos);
-            for (let pos of path.path) {
+            for (const pos of path.path) {
                 if (Game.map.getTerrainAt(pos) === "swamp") {
                     const constructionSites = pos.lookFor(LOOK_CONSTRUCTION_SITES) as ConstructionSite[];
                     if (constructionSites.length === 0) {
