@@ -1,6 +1,6 @@
+import { addProcess, getProcessById } from "../../kernel/kernel/kernel";
 import Process = require("../../kernel/kernel/process");
 import { ProcessPriority } from "../constants";
-import { addProcess, getProcessById } from "../../kernel/kernel/kernel";
 interface CreepRequest {
     pid: number;
     creepID: string;
@@ -22,53 +22,53 @@ class SpawnProcess extends Process {
         return this.memory.roomName;
     }
     public spawn(id: string, bodyMap: bodyMap, pid: number, priority = 10): number {
-        let memory = this.memory;
-        let existing = _.find(<CreepRequest[]>memory.requestList, r => (r.creepID === id) && (r.pid === pid));
+        const memory = this.memory;
+        const existing = _.find(memory.requestList as CreepRequest[], (r) => (r.creepID === id) && (r.pid === pid));
         if (existing) {
             return -1;
         } else {
-            let creepRequest = { pid: pid, creepID: id, bodyParts: bodyMap, priority: priority };
+            const creepRequest = { pid, creepID: id, bodyParts: bodyMap, priority };
             memory.requestList.push(creepRequest);
             return 0;
         }
 
     }
     public run(): number {
-        let colonyProcess = getProcessById(this.parentPID);
+        const colonyProcess = getProcessById(this.parentPID);
         if (!colonyProcess)
             return this.stop(0);
 
-        let roomName = this.memory.roomName;
+        const roomName = this.memory.roomName;
         if (!roomName)
             return this.stop(0);
-        const makeBody = function (bodyMap: bodyMap): string[] {
+        const makeBody = function(bodyMap: bodyMap): string[] {
             // TODO : Fix the part map, this need to include all part type somehow
-            let partMap: { [s: string]: string } = {
-                WORK: WORK,
-                MOVE: MOVE,
-                CARRY: CARRY,
-                ATTACK: ATTACK,
-                CLAIM: CLAIM,
+            const partMap: { [s: string]: string } = {
+                WORK,
+                MOVE,
+                CARRY,
+                ATTACK,
+                CLAIM,
             };
-            let replicatePart = function (times: number, part: string) {
-                return _.map(_.times(times, x => x),
+            const replicatePart = function(times: number, part: string) {
+                return _.map(_.times(times, (x) => x),
                     () => partMap[part]);
             };
-            return <string[]>_.chain(bodyMap).map(replicatePart).flatten().value();
+            return _.chain(bodyMap).map(replicatePart).flatten().value() as string[];
         };
 
-        let memory = this.memory;
+        const memory = this.memory;
         memory.requestList = memory.requestList || [];
-        memory.requestList = _.sortBy(<CreepRequest[]>memory.requestList, i => i.priority);
-        let request: CreepRequest = memory.requestList.pop();
-        let spawn = this.findFreeSpawn(this.memory.roomName);
+        memory.requestList = _.sortBy(memory.requestList as CreepRequest[], (i) => i.priority);
+        const request: CreepRequest = memory.requestList.pop();
+        const spawn = this.findFreeSpawn(this.memory.roomName);
         if (request) {
             if (spawn) {
                 const body = makeBody(request.bodyParts);
                 const canSpawn = spawn.canCreateCreep(body);
                 if (canSpawn === OK) {
-                    let process: any = getProcessById(request.pid);
-                    let creepName = spawn.createCreep(body);
+                    const process: any = getProcessById(request.pid);
+                    const creepName = spawn.createCreep(body);
                     process.receiveCreep(request.creepID, Game.creeps[creepName]);
                 }
             }
@@ -78,8 +78,8 @@ class SpawnProcess extends Process {
     }
 
     private findFreeSpawn(roomName: string) {
-        let spawns = _.filter(Game.spawns,
-            function (spawn) {
+        const spawns = _.filter(Game.spawns,
+            function(spawn) {
                 return spawn.room.name === roomName &&
                     spawn.spawning == null &&
                     spawn.canCreateCreep([MOVE]) === OK &&
