@@ -1,6 +1,6 @@
+import { addProcess, getProcessById, sleepProcess } from "../../kernel/kernel/kernel";
 import Process = require("../../kernel/kernel/process");
 import { getSpawnProcess } from "../../utils/spawn";
-import { sleepProcess, addProcess, getProcessById } from "../../kernel/kernel/kernel";
 interface UpgraderMemory {
     name: string | undefined;
     hasCourier: boolean;
@@ -12,7 +12,7 @@ interface UpgraderMemory {
 class UpgraderProcess extends Process {
     public memory: UpgraderMemory;
     public static start(roomName: string, parentPID: number) {
-        let p = new UpgraderProcess(0, parentPID);
+        const p = new UpgraderProcess(0, parentPID);
         addProcess(p);
         p.memory.roomName = roomName;
         return p.pid;
@@ -22,7 +22,7 @@ class UpgraderProcess extends Process {
         return "components.processes.room.upgrader";
     }
     public run(): number {
-        let colonyProcess = getProcessById(this.parentPID);
+        const colonyProcess = getProcessById(this.parentPID);
         if (!colonyProcess)
             return this.stop(0);
 
@@ -31,14 +31,13 @@ class UpgraderProcess extends Process {
             sleepProcess(this, 1500);
             return 0;
         }
-        let memory = this.memory;
-        let upgraderName = memory.name;
+        const memory = this.memory;
+        const upgraderName = memory.name;
         if (upgraderName && Game.creeps[upgraderName]) {
             this.runCreep(upgraderName);
         } else {
             this.spawnCreep();
         }
-
 
         if (this.memory.hasCourier) {
             const courierName = this.memory.courierName;
@@ -51,9 +50,9 @@ class UpgraderProcess extends Process {
     }
 
     public runCreep(creepName: string): number {
-        let creep = Game.creeps[creepName];
+        const creep = Game.creeps[creepName];
         let storage: StructureStorage | StructureContainer | null = Game.getObjectById(this.memory.containerID!) as StructureStorage | StructureContainer;
-        let controller = creep.room.controller!;
+        const controller = creep.room.controller!;
         if (!storage)
             storage = this.getEnergySource();
         if (!storage) {
@@ -69,7 +68,7 @@ class UpgraderProcess extends Process {
             creep.withdraw(storage, RESOURCE_ENERGY);
             this.work(creep);
         } else {
-            let pos = _.filter(storage.pos.adjacentPositions(), pos => pos.inRangeTo(controller.pos, 3));
+            const pos = _.filter(storage.pos.adjacentPositions(), (pos) => pos.inRangeTo(controller.pos, 3));
             if (pos.length > 0)
                 creep.moveTo(pos[0]);
             else {
@@ -83,7 +82,7 @@ class UpgraderProcess extends Process {
 
     private runCourier(creepName: string): number {
         const creep = Game.creeps[creepName];
-        let storage: StructureStorage | StructureContainer | null = Game.getObjectById(this.memory.containerID!) as StructureStorage | StructureContainer;
+        const storage: StructureStorage | StructureContainer | null = Game.getObjectById(this.memory.containerID!) as StructureStorage | StructureContainer;
         if (!storage) {
             return 0;
         }
@@ -97,7 +96,6 @@ class UpgraderProcess extends Process {
             creep.transfer(storage, RESOURCE_ENERGY);
         }
         return 0;
-
 
     }
     public receiveCreep(id: string, creep: Creep) {
@@ -113,7 +111,7 @@ class UpgraderProcess extends Process {
         const controller = room.controller!;
         if (!storage)
             return null;
-        let pos = _.filter(storage.pos.adjacentPositions(), pos => pos.inRangeTo(controller.pos, 3));
+        const pos = _.filter(storage.pos.adjacentPositions(), (pos) => pos.inRangeTo(controller.pos, 3));
         if (pos.length === 0) {
             const allContainers = room.find(FIND_STRUCTURES, { filter: (s: Structure) => s.structureType === STRUCTURE_CONTAINER && s.pos.inRangeTo(controller.pos, 3) }) as StructureContainer[];
             if (allContainers.length > 0) {
@@ -142,11 +140,11 @@ class UpgraderProcess extends Process {
             creep.upgradeController(creep.room.controller!);
     }
     private spawnCreep() {
-        let roomName = this.memory.roomName;
-        let spawnProcess = getSpawnProcess(this.memory.roomName);
+        const roomName = this.memory.roomName;
+        const spawnProcess = getSpawnProcess(this.memory.roomName);
         let energy = Game.rooms[roomName].energyCapacityAvailable;
         energy = energy - 50;
-        let multiplier = Math.floor(energy / 250);
+        const multiplier = Math.floor(energy / 250);
         if (spawnProcess) {
             spawnProcess.spawn("upgrader", { CARRY: 1, WORK: 2 * multiplier, MOVE: multiplier },
                 this.pid);
@@ -154,10 +152,10 @@ class UpgraderProcess extends Process {
     }
 
     private spawnCourier() {
-        let roomName = this.memory.roomName;
-        let spawnProcess = getSpawnProcess(this.memory.roomName);
-        let energy = Game.rooms[roomName].energyCapacityAvailable;
-        let multiplier = Math.floor(energy / 100);
+        const roomName = this.memory.roomName;
+        const spawnProcess = getSpawnProcess(this.memory.roomName);
+        const energy = Game.rooms[roomName].energyCapacityAvailable;
+        const multiplier = Math.floor(energy / 100);
         if (spawnProcess) {
             spawnProcess.spawn("courier", { CARRY: multiplier, MOVE: multiplier },
                 this.pid);
